requirements.txt
aioredis==1.3.1
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backcall==0.2.0
black==21.12b0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.3
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.1
Django==4.0.1
django-browser-reload==1.3.0
django-crispy-forms==1.14.0
django-entangled==0.4
django-extensions==3.1.5
django-tempus-dominus==5.1.2.16
executing==0.8.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
msgpack==1.0.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==1.2.3
traitlets==5.1.1
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
wcwidth==0.2.5
zope.interface==5.4.0
